test: clean up more integration tests in pop-chain
